feat(r-date-range): rename date to formatDate
