+++ packages/recomponents/src/common/helpers/index.js
@@ -1,13 +1,13 @@
-import date from './date';
 import debounce from './debounce';
 import deepClone from './deep-clone';
+import formatDate from './format-date';
 import kebabCase from './kebab-case';
 import isObject from './is-object';
 
 export default {
-    date,
     debounce,
     deepClone,
+    formatDate,
     kebabCase,
     isObject,
 };
